simplify word lists
each diceware word list consists of lines of words. each word is
prefixed (exception of the rule: diceware8k) with a "dicechain", a
string which represents the result of a dice throw: '1' when the dice
shows '1', a '5' when the dice shows a '5' and so on.

the used word lists are compiled into words_xyz.go. because each list is
roughly 80/90kbyte in size, my initial plan was to gzip the raw lists,
base64 it again and use this internally. the result was something in the
ballpark of 50kbyte, but the lists could not be checked visually anymore.

by removing the prefix the lists shrink to ~50kbyte and can be checked
visually. the prefix for each position in the list can be calculated by
transforming the index from base-10 to base-6 ('0'-'5') and adjusting the
shown value by '1' ('0' becomes '1', '5' becomes '6' etc). to convert
a "dicechain" of the user to an index in the word list, the values are
first decreased by 1 ('6' becomes '5' etc) and then converting a base-6
number to base-10:

  dicechain     base10
      11111 <-> 0000
      11112 <-> 0001
      11116 <-> 0005
      12345 <-> 0310
      54321 <-> 5910
      66666 <-> 7775