combine -horizontal and -dump
grouping the dumped word lists into 6 columns looks pretty nice and
brings down the number of lines down to 1296 (from 7776). this makes it
much easier to print the lists out to dead wood ("paper").

the -horizontal flag is available already and it seems to be a good fit.
so, -horizontal affects the output of -dump now as well.